+++ CHANGELOG.md
@@ -32,6 +32,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * safer no longer uses `temp_file` on windows
 * Erroneous assertion errors in reproject and kwcoco-to-geotiffs
 * Issue where DINO_SV would write region models to the out-site-manifest, now correctly points at site models.
+* Bug in coco-align where nodata values were not properly set on data that moved through gdal-merge.
 
 
 ## Version 0.5.6 - Target 2023-04-30